文章支持标签分类

+++ src/main/java/com/bzw875/blog/repository/TagRepository.java
@@ -1,13 +1,11 @@
 package com.bzw875.blog.repository;
 
 
-import org.springframework.data.repository.CrudRepository;
-
 import com.bzw875.blog.model.Tag;
+import org.springframework.data.repository.CrudRepository;
 
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface TagRepository extends CrudRepository<Tag, Long> {
-
+public interface TagRepository extends CrudRepository<Tag, Integer> {
 }
